+++ src/main/java/svenhjol/charm/base/integration/package-info.java
@@ -1,0 +1,3 @@
+@javax.annotation.ParametersAreNonnullByDefault
+@mcp.MethodsReturnNonnullByDefault
+package svenhjol.charm.base.integration;